Refactored Round case class into its own file
